Neighbourhood verification for ensembles

Neighbourhood verification for ensembles

nbhd_verify(
  .fcst,
  obs,
  threshold,
  radius,
  comparator = "ge",
  include_low = TRUE,
  include_high = TRUE,
  ...
)

Arguments

threshold

A numeric vector of thresholds to for which to compute the neighbourhood scores. Can also be a character vector with numeric values between 0 and 1 preceded by a "q" to denote quantile thresholds.

radius

The radius, in pixels of the neighbourhoods for which to compute the scores.

.data

A "harp_ens_grid_df" data frame or a geolist.

obs_col

The name of the observations column in the "harp_ens_grid_df", or if .data is a geolist, a geofield.