Run spatial verification on a (for now) deterministic forecast

Run spatial verification on a (for now) deterministic forecast

Run spatial verification on a (for now) deterministic forecast

verify_spatial(
  dttm,
  start_date = NULL,
  end_date = NULL,
  by = NULL,
  parameter,
  fcst_model = harpSpatial_conf$fcst_model,
  lead_time = harpSpatial_conf$lead_time,
  lt_unit = harpSpatial_conf$lt_unit,
  scores = harpSpatial_conf$scores,
  members = harpSpatial_conf$members,
  fc_file_path = harpSpatial_conf$fc_file_path,
  fc_file_template = harpSpatial_conf$fc_file_template,
  fc_file_format = harpSpatial_conf$fc_file_format,
  fc_file_opts = harpSpatial_conf$fc_file_opts,
  fc_domain = harpSpatial_conf$fc_domain,
  fc_interp_method = harpSpatial_conf$fc_interp_method,
  fc_accumulation = harpSpatial_conf$fc_accumulation,
  ob_file_path = harpSpatial_conf$ob_file_path,
  ob_file_template = harpSpatial_conf$ob_file_template,
  ob_file_format = harpSpatial_conf$ob_file_format,
  ob_file_opts = harpSpatial_conf$ob_file_opts,
  ob_domain = harpSpatial_conf$ob_domain,
  ob_interp_method = harpSpatial_conf$ob_interp_method,
  ob_accumulation = harpSpatial_conf$ob_accumulation,
  verif_domain = harpSpatial_conf$verif_domain,
  use_mask = harpSpatial_conf$use_mask,
  window_sizes = harpSpatial_conf$window_sizes,
  thresholds = harpSpatial_conf$thresholds,
  sqlite_path = harpSpatial_conf$sqlite_path,
  sqlite_file = harpSpatial_conf$sqlite_file,
  return_data = FALSE
)

verify_spatial(
  dttm,
  start_date = NULL,
  end_date = NULL,
  by = NULL,
  parameter,
  fcst_model = harpSpatial_conf$fcst_model,
  lead_time = harpSpatial_conf$lead_time,
  lt_unit = harpSpatial_conf$lt_unit,
  scores = harpSpatial_conf$scores,
  members = harpSpatial_conf$members,
  fc_file_path = harpSpatial_conf$fc_file_path,
  fc_file_template = harpSpatial_conf$fc_file_template,
  fc_file_format = harpSpatial_conf$fc_file_format,
  fc_file_opts = harpSpatial_conf$fc_file_opts,
  fc_domain = harpSpatial_conf$fc_domain,
  fc_interp_method = harpSpatial_conf$fc_interp_method,
  fc_accumulation = harpSpatial_conf$fc_accumulation,
  ob_file_path = harpSpatial_conf$ob_file_path,
  ob_file_template = harpSpatial_conf$ob_file_template,
  ob_file_format = harpSpatial_conf$ob_file_format,
  ob_file_opts = harpSpatial_conf$ob_file_opts,
  ob_domain = harpSpatial_conf$ob_domain,
  ob_interp_method = harpSpatial_conf$ob_interp_method,
  ob_accumulation = harpSpatial_conf$ob_accumulation,
  verif_domain = harpSpatial_conf$verif_domain,
  use_mask = harpSpatial_conf$use_mask,
  window_sizes = harpSpatial_conf$window_sizes,
  thresholds = harpSpatial_conf$thresholds,
  sqlite_path = harpSpatial_conf$sqlite_path,
  sqlite_file = harpSpatial_conf$sqlite_file,
  return_data = FALSE
)

Arguments

dttm

A vector of date time strings to read. Can be in YYYYMMDD, YYYYMMDDhh, YYYYMMDDhhmm, or YYYYMMDDhhmmss format. Can be numeric or character. seq_dttm can be used to generate a vector of equally spaced date-time strings.

start_date

Date of the first forecast to read.

end_date

Date of the last forecast to read.

by

The time between forecasts. Should be a string of a number followed by a letter, where the letter gives the units - may be "d" for days, "h" for hours or "m" for minutes.

parameter

The parameters to read as a character vector.

fcst_model

The name of the (deterministic or EPS) model.

lead_time

The lead times to read as a numeric vector. Should be in the units that are also used in fc_file_template.

lt_unit

The unit used for lead_time. Can be "h" (hours), "m" (minutes), "s" (seconds)

members

The (numbers of the) ensemble members to read. While Netcdf and grib2 files can contain multiple members, for other formats we assume they are in separate files (see also fc_file_template)

fc_file_path

The top level path for the forecast files to read.

fc_file_template

The file type to generate the template for. Can be "harmoneps_grib", "harmeoneps_grib_fp", "harmoneps_grib_sfx", "meps_met", "harmonie_grib", "harmonie_grib_fp", "harmone_grib_sfx", "vfld", "vobs", or "fctable". If anything else is passed, it is returned unmodified. In this case substitutions can be used. Available substitutions are {YYYY} for year, {MM} for 2 digit month with leading zero, {M} for month with no leading zero, and similarly {DD} or {D} for day, {HH} or {H} for hour, {mm} or {m} for minute. Also {LDTx} for lead time and {MBRx} for ensemble member where x is the length of the string including leading zeros - can be omitted or 2, 3 or 4. Note that the full path to the file will always be file_path/template.

fc_file_format

The format of the files to read. Can be e.g. "fa" or "grib".

fc_file_opts

A list with format-specific options for the reader function.

fc_domain

The forecast domain. If provided, the fc reading can be made faster by not extracting domain information (format option meta).

fc_interp_method

Interpolation method to be used when transforming a forecast field to the verification grid.

fc_accumulation

The accumulation type of the forecast. This is only used for accumulated parameters (e.g. precipitation). NULL signifies that the field is accumulated from the start of the model run. Otherwise this should be a string containing a numerical value and a time unit, e.g. "15m" or "1h".

ob_file_path

The top level path for the forecast files to read.

ob_file_template

The file type to generate the template for. Can be "harmoneps_grib", "harmeoneps_grib_fp", "harmoneps_grib_sfx", "meps_met", "harmonie_grib", "harmonie_grib_fp", "harmone_grib_sfx", "vfld", "vobs", or "fctable". If anything else is passed, it is returned unmodified. In this case substitutions can be used. Available substitutions are YYYY for year, {MM} for 2 digit month with leading zero, {M} for month with no leading zero, and similarly {DD} or {D} for day, {HH} or {H} for hour, {mm} or {m} for minute. Also {LDTx} for lead time and {MBRx} for ensemble member where x is the length of the string including leading zeros - can be omitted or 2, 3 or 4. Note that the full path to the file will always be file_path/template.

ob_file_format

The format of the files to read. Can be e.g. "hdf5" or "grib".

ob_file_opts

A list with format-specific options for the reader function.

ob_domain

The observation domain. If provided, the obs reading can be made faster by not extracting domain information (format option meta).

ob_interp_method

Interpolation method to be used when transforming a forecast field to the verification grid.

ob_accumulation

The accumulation type of the observation (or reference). This is only used for accumulated parameters (e.g. precipitation). NULL signifies that the field is accumulated from the start of the model run. That is probably rare for observations. Otherwise this should be a string containing a numerical value and a time unit, e.g. "15m" or "1h". "0" means an instantaneous value.

verif_domain

A geodomain that defines the common verification grid.

window_sizes

Scales used for fuzzy methods like FSS. A vector of box sizes. All values must be odd integers (so the central point is really in the center of a box).

thresholds

Thresholds used for FSS, ...

sqlite_path

If specified, SQLite files are generated and written to this directory.

sqlite_file

Name of SQLite file.

return_data

If TRUE, the result is returned as a list of tables.

model

The name of the (deterministic or EPS) model.

fc_options

A list with format-specific options for the reader function.

ob_options

A list with format-specific options for the reader function.

start_date,

end_date, by `r lifecycle::badge("deprecated")` Date of the first and last forecast to read. Please use `dttm` together with seq_dttm to generate equally spaced date-times.

...

Not used at thispoint (more info to be added).

Value

A list containting tibbles for all scores.

A list containting tibbles for all scores.